Clear out SettingsWindow with content in other pages
